+++ CHANGELOG.md
@@ -3,6 +3,30 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
+### Added
+  - Map attributes now accepts dynamic fields, if corresponding attribute is
+included in brackets
+
+### Changed
+  - EEG:BrainVision: removed `PatientId`, `SessionId` and `RecordingNumber`
+from attributes
+  - EEG:BrainVision: functions `getSubId` and `getSesId` returns `None`
+  - If subject Id and session Id is set to None, the corresponding values
+are extracted from file name, assumming it is formatted following BIDS 
+(fields `sub` and `ses` respectively)
+  - If map attributes dictionary is empty, it will match all recordings
+  - BidsSession: BidsSession class can be initialized with subject and 
+session values. If these values are set, corresponding attributes will 
+be locked
+  - Preparation: if subject or session values are `None`, they are calculated
+for each recording in serie
+
+### Fixed
+  - EEG:BrainVision: fixed `_loadFile` function name
+  - EEG:BrainVision: added `acq` to template map
+  - EEG:BrainVision: Fixed TaskName and SamplingFrequency json fields
+
+
 ## [1.0.1] - 2020-04-20
 
 ### Changed